Update downloader to use Summary by Halves instead of Box Score
Some of the pages now return empty box scores, although the data is
clearly there. The Summary by Halves page has everything except the
number of minutes played, but that's ok since it's not used anyway.

The changes to the XSL are fairly hack-and-slash to get it working
again. At some point that should be cleaned up.