Update downloader to use Box Score instead of Summary by Halves
Previously I had updated to use the Summary by Halves page because
some of them came back blank. That bug seems to have been fixed in
the last 7 years. Now I need to revert back to doing that because
the GameId is no longer present in the game-by-game page, and it's
much simpler to avoid doing the game-by-game indirection altogether
and start using the "contest ID" instead of the "game ID".